+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+* Upgrade to Pixelfed v0.11.9
+* Add ActivityPubEnabled parameter
+* Enable Mastodon Login feature
+* Add health check URI
+
 # 1.0.0
 
 * Initial development